split book exporter into modules
